requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/development.txt requirements/development.in
#
asn1crypto==0.24.0
atomicwrites==1.3.0       # via pytest
attrs==18.2.0             # via pytest
cffi==1.11.5
click==7.0
cryptography==2.5
filelock==3.0.10          # via tox
more-itertools==5.0.0     # via pytest
pip-tools==3.3.2
pluggy==0.8.1             # via pytest, tox
py==1.7.0                 # via pytest, tox
pycparser==2.19
pytest==4.2.0
pyyaml==3.13
six==1.12.0
toml==0.10.0              # via tox
tox==3.7.0
virtualenv==16.3.0        # via tox
